Refactor log module
The goal of this commit is to make better use of Python's logging
module for debugging as well as user feedback.

Before, in-toto used wrappers for logging functions directly on the
in_toto.log module.

Now these wrappers are removed and instead we register a base
in_toto logger, using a custom subclass with an enhanced error
method and a convenience method to set the logging level, e.g. from
command line interfaces.

Read in_toto.log's doc header for detailed usage instructions.

The commit also removes LOG_LEVEL setting and adds a DEBUG setting
(boolean) which affects the format and log level of the new in-toto
base logger.

The commit also replaces some warning with info statements (for
user feedback we should always use info).