Add optional short_keyid to gpg signature schema
The field can be used for the short gpg fingerprint as
contained in the gpg signature type 16 issuer subpacket
(see rfc4880 5.2.3.1).

However, the short fingerprint should not be used for signature
verification, as keys and signatures must be matched by their
full keyids.

The short keyid should only be used determine the full keyid.