Update gpg signature subpacket parsing
Before we looked for short keyids only in unhashed subpackets and
for long keyids only in hashed subpackets and considered the first
occurrence of a subpacket of a given type only.

This commit updates parsing to look for any gpg signature subpacket
in both the hashed and unhashed subpacket sections of a signature,
and to consider only the last subpacket, which conforms with the
RFC. Furthermore, hashed subpackets are given precedence over
unhashed packets.

See code comments for more details and RFC references.

The commit also updates the error message and adds an additional
check that short and long keyids must match if both are specified.

NOTES:
The error checks are currently excluded from coverage, since
triggering them would require manually crafting partially invalid
gpg signature data.

It appears that this update only affects very rare corner cases.