Add optional info to signature
Update parse_signature_packet to add an optional opaque info
dictionary to the returned signature. Whether it is added or
not may be configured via boolean "include_info" parameter. Per
default it is not added.

Currently the info dictionary contains the signature_type and
hash_algorithm and an empty list that may be used to add
subpackets such as a key expiration date.
The corresponding subpacket parsing code is already there
(commented out).