Make gpg KeyExpirationError message timezone aware
As per RFC4880 section 3.5 time fields, such as the key creation
time contain the number of seconds elapsed since midnight, 1
January 1970 UTC.

This commit modifies the error message in a KeyExpirationError
to use timezone UTC for creation and calculated expiration date
and adds the timezone specifier to the  date trings.

The timezone name, i.e. "UTC" is hardcoded, because the %Z
specifier uses different names on different python versions.

Expiration assessment is unaffected, as it uses time.time(), which
is compatible with creation time (i.e. seconds since epoch).