Add MANIFEST.in to include files in source distro
MANIFEST.in is used to specify files (patterns) to be added to
the source distribution (e.g. on `python setup.py sdist`) in
addition to files that are added by default.

This commit configures MANIFEST.in to add files that are
important/helpful for downstream packagers, i.e. readme, license,
tests and test data and config.

Kudos to @koobs for bringing this to our attention.