Make aggregate_tests callable from project dir
This is Python testing convention and allows coverage to correctly
record the paths of the source files and left-strip the parts
leading to the project directory.

This commit also replaces manual test loading with the unittest
`discover` function, which adds some path magic, so that tests
are actually run over the source code in the project dir.

For the sake of simplified test discovery, this commit removes
randomization of test module loading.
It also does not seem beneficial to just shuffle the test
modules and not the test cases or test methods.