Use hash algo per ecdsa curve and add nistp384
This commit adds support for verifying ecdsa signature on the
nistp384 curve with sha384 digests to the internal ecdsa_keys
module.

It does so by adding module global helper dictionary to map schemes
to hash algorithms.

Note: This commit tries to blend in with the current sslib design.
In future work we should:
- define securesystemslib-wide constants instead of hardcoding
  strings over and over again (see item 3 in #183)