Remove support for gzipped files in load_json_file
Compression related functions were removed from securesystemslib as it was
deemed more appropriate for callers to handle compression/decompression at
a different level. However, the automatic decompression functionality in
load_json_file() had remained.

Remove that change and tidy up load_json_file() to only close the
fileobject once. Since Python 2.5 a finally block is always called,
regardless of whether there is an exception or not. Therefore there's no
need to close the fileobject in the else block and the finally block.

Signed-off-by: Joshua Lock <jlock@vmware.com>