Split up util.get_file_details()
The util.get_file_details() function returns the length and hashes
of the specified file. This makes a lot of sense for generating
information about TUF targets, where the length and hashes of
each target are required, but is potentially too costly function
when generating information about TUF metadata files for the
snapshot role, where lengths and hashes are optional.
Snapshot metadata may only want hashes, or only want lengths,
and we don't want to force an adopter to implement their own subset
of the functionality in get_file_details().

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>