Lower strict coverage threshold
Our coverage is currently very close to the strict 99 threshold,
and we sometimes go below due to different rounding strategies on
different builds (different coverage versions?) or if we add/remove
code and the relative coverage ratio changes slightly.

This commit lowers the strict coverage threshold from 99 to 97.

are very close to that threshold