Add and adopt decryption password interface helper
Add helper for private key import interface functions to use the,
passed decryption password, or to prompt for a password based on a
passed 'prompt' boolean, or to treat the key as unencrypted.

The helper aggregates and replaces repetitive code in
'interface.import_{rsa, ed25519, ecdsa}_privatekey_from_file'
functions, to make the password handling consistent across these
functions.

This commit only adopts the helper for rsa and ed25519 import
functions. The ecdsa function requires more invasive changes (see
subsequent commit).

**Change of behavior**:
Passing an empty string for a password no longer raises a
FormatError, instead it is left to the underlying decryption
function to fail with something like a "wrong password error",
because there shouldn't be a difference between a wrong and a wrong
empty password.

See test changes for change of behavior.