Catch and reraise cryptography error
When calling securesystemslib.keys.verify_signature() and consequently
call securesystemslib.keys.verify_signature() with a malformed
public key (it has a raw bytes encoded in HEX instead of PEM) then
we get a ValueError from cryptography with message:
"Could not deserialize key data. The data may be in an incorrect
format, it may be encrypted with an unsupported algorithm..."

It's good if we replace that ValueError with
securesystemslib.FormatError in order to be consistent.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>