Replace deprecated distutils.version.StrictVersion.
distutils is deprecated as of Python 3.10. We were only using it for
`StrictVersion`, which parses a version string and lets you order the
result.

The [official recommendation] from the packaging team is to replace
`StrictVersion` with a dependency on the `packaging` package, but:

1. The Python version parser was a little different from GPG versioning
   anyway because it's made for Python versions (it had support for
   "prerelease tags").
2. I wanted to avoid adding dependencies if possible.

[official recommendation]: https://github.com/pypa/packaging/issues/520

This PR adds a new `Version` class in `securesystemslib.gpg.util` and
uses that instead.

Signed-off-by: Zachary Newman <z@znewman.net>