signer: add pkcs11 lib pseudo-singleton
Add ugly helper function to not load pkcs11 library more than once,
while allowing to load it explicitly.

This important for tests, where we need to setup a softhsm config
file before we load the library, and need to load the library
before we use the Signer.

**Alternative approachs**:
Load and unload for each use, just like we create and close session
for each use:
- unload is implemented on `PyKCS11Lib.__del__`
- performance?

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>