signer: fix imports in sigstore signer
The sigstore library is an optional and circular dependency in
securesystemslib. Importing it in the methods, where it is needed,
allows importing the module from within sigstore (circular import),
or if sigstore is not installed.

This commit wraps the local scope imports in try/except to fail
gracefully if the methods are called w/o sigstore available, and
locally disables the related linter warning.

The commit also disables a linter import warning in the related
test, which requires sigstore. Alternatively, we could install
sigstore in the lint job, which does not seem worth the cycles,
just for one method call.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>